Changed headers to HPP format
