+++ include/MINX/Graphics/Color.hpp
@@ -23,11 +23,12 @@ freely, subject to the following restrictions:
 > 3\. This notice may not be removed or altered from any source
 > distribution.
 >
-        */
-#ifndef COLOR_H_
-#define COLOR_H_
+*/
 
-#include "../API.h"
+#ifndef MINX_COLOR_HPP_
+#define MINX_COLOR_HPP_
+
+#include "../API.hpp"
 
 namespace MINX
 {
@@ -40,85 +41,112 @@ namespace MINX
 			/** The red component of the color.
 			 */
 			double R;
+
 			/** The green component of the color.
 			 */
 			double G;
+
 			/** The blue component of the color.
 			 */
 			double B;
+
 			/** The alpha component of the color.
 			 */
 			double A;
+
 			/** Constructs the Color with a default value of White
 			 */
 			Color();
+
 			/** Constructs a color with the given red, green, blue, and alpha values.
 			 */
 			Color(double r, double g, double b, double a);
+
 			/** Constructs a color with the given red, green, and blue values.
 			 */
 			Color(double r, double g, double b);
+
 			/** Tests equality between two Colors.
 			 */
 			bool operator==(const Color& compareTo);
+
 			/** Tests inequality between two Colors
 			 */
 			bool operator!=(const Color& compareTo);
+
 			/** Tests if a color is lighter than another Color
 			 */
 			bool operator>(const Color& compareTo);
+
 			/** Tests if a color is lighter than or the same as another Color
 			 */
 			bool operator>=(const Color& compareTo);
+
 			/** Tests if a Color is darker than another Color
 			 */
 			bool operator<(const Color& compareTo);
+
 			/** Tests if a Color is darker than or the same as another Color
 			 */
 			bool operator<=(const Color& compareTo);
+
 			/** Inverts a Color
 			 */
 			Color operator!()const;
+
 			/** Adds 2 Colors
 			 */
 			Color operator+(const Color& addTo);
+
 			/** Adds 2 Colors
 			 */
 			Color operator+=(const Color& addTo);
+
 			/** Subtracts a color from another color
 			 */
 			Color operator-(const Color& subtract);
+
 			/** Subtracts a color from another color
 			 */
 			Color operator-=(const Color& subtract);
+
 			/** Multiplies 2 colors
 			 */
 			Color operator*(const Color& multiplyBy);
+
 			/** Multiplies 2 colors
 			 */
 			Color operator*=(const Color& multiplyBy);
+
 			/** Divides a color by another color
 			 */
 			Color operator/(const Color& divideBy);
+
 			/** Divides a color by another color
 			 */
 			Color operator/=(const Color& divideBy);
+
 			/** Cross multiplies 2 colors
 			 */
 			Color Cross(const Color& crossWith);
+
 			/** Multiplies a color by a scale
 			 */
 			Color operator*(const double& scale);
+
 			/** Multiplies a color by a scale
 			 */
 			Color operator*=(const double& scale);
+
 			/** Divides a color by a scale
 			 */
 			Color operator/(const double& scale);
+
 			/** Divides a color by a scale
 			 */
 			Color operator/=(const double& scale);
-			/** \defgroup Named Colors
+
+			/** \defgroup colors Named Colors
 			 * @{
 			 */
 			/** \defgroup Transparency
@@ -126,8 +154,9 @@ namespace MINX
 			 */
 			static Color TransparentBlack;
 			static Color TransparentWhite;
+
 			/** @}*/
-			/** \defgroup Shades of Pink
+			/** \defgroup Pink Shades of Pink
 			 * @{
 			 */
 			static Color Pink;
@@ -136,8 +165,9 @@ namespace MINX
 			static Color DeepPink;
 			static Color PaleVioletRed;
 			static Color MediumVioletRed;
+
 			/** @}*/
-			/** \defgroup Shades of Red
+			/** \defgroup Red Shades of Red
 			 * @{
 			 */
 			static Color LightSalmon;
@@ -149,8 +179,9 @@ namespace MINX
 			static Color FireBrick;
 			static Color DarkRed;
 			static Color Red;
+
 			/** @}*/
-			/** \defgroup Shades of Orange
+			/** \defgroup Orange Shades of Orange
 			 * @{
 			 */
 			static Color OrangeRed;
@@ -158,8 +189,9 @@ namespace MINX
 			static Color Coral;
 			static Color DarkOrange;
 			static Color Orange;
+
 			/** @}*/
-			/** \defgroup Shades of Yellow
+			/** \defgroup Yellow Shades of Yellow
 			 * @{
 			 */
 			static Color Yellow;
@@ -173,8 +205,9 @@ namespace MINX
 			static Color Khaki;
 			static Color DarkKhaki;
 			static Color Gold;
+
 			/** @}*/
-			/** \defgroup Shades of Brown
+			/** \defgroup Brown Shades of Brown
 			 * @{
 			 */
 			static Color Cornsilk;
@@ -194,8 +227,9 @@ namespace MINX
 			static Color Sienna;
 			static Color Brown;
 			static Color Maroon;
+
 			/** @}*/
-			/** \defgroup Shades of Green
+			/** \defgroup Green Shades of Green
 			 * @{
 			 */
 			static Color DarkOliveGreen;
@@ -218,8 +252,9 @@ namespace MINX
 			static Color ForestGreen;
 			static Color WebGreen;
 			static Color DarkGreen;
+
 			/** @}*/
-			/** \defgroup Shades of Cyan
+			/** \defgroup Cyan Shades of Cyan
 			 * @{
 			 */
 			static Color MediumAquamarine;
@@ -234,8 +269,9 @@ namespace MINX
 			static Color CadetBlue;
 			static Color DarkCyan;
 			static Color Teal;
+
 			/** @}*/
-			/** \defgroup Shades of Blue
+			/** \defgroup Blue Shades of Blue
 			 * @{
 			 */
 			static Color LightSteelBlue;
@@ -253,8 +289,9 @@ namespace MINX
 			static Color DarkBlue;
 			static Color Navy;
 			static Color MidnightBlue;
+
 			/** @}*/
-			/** \defgroup Shades of Purple
+			/** \defgroup Purple Shades of Purple
 			 * @{
 			 */
 			static Color Lavender;
@@ -274,8 +311,9 @@ namespace MINX
 			static Color DarkSlateBlue;
 			static Color SlateBlue;
 			static Color MediumSlateBlue;
+
 			/** @}*/
-			/** \defgroup Shades of White
+			/** \defgroup White Shades of White
 			 * @{
 			 */
 			static Color White;
@@ -295,8 +333,9 @@ namespace MINX
 			static Color Linen;
 			static Color LavenderBlush;
 			static Color MistyRose;
+
 			/** @}*/
-			/** \defgroup Shades of Gray
+			/** \defgroup Gray Shades of Gray
 			 * @{
 			 */
 			static Color Gainsboro;
@@ -308,11 +347,13 @@ namespace MINX
 			static Color LightSlateGray;
 			static Color SlateGray;
 			static Color DarkSlateGray;
+
 			/** @}*/
 			/** \defgroup Black
 			 * @{
 			 */
 			static Color Black;
+
 			/** @}*/
 			/** @}*/
 		};
