+++ include/MINX/Graphics/ShaderFactory.hpp
@@ -1,4 +1,3 @@
-
 /*
 # MINX
 
@@ -23,12 +22,12 @@ freely, subject to the following restrictions:
 > 3\. This notice may not be removed or altered from any source
 > distribution.
 >
-        */
+*/
 
-#ifndef SHADER_FACTORY_H_
-#define SHADER_FACTORY_H_
+#ifndef MINX_SHADERFACTORY_HPP_
+#define MINX_SHADERFACTORY_HPP_
 
-#include "../API.h"
+#include "../API.hpp"
 
 #include <fstream>
 #include <vector>
