+++ include/MINX/Input/IGenericHID.hpp
@@ -1,4 +1,3 @@
-
 /*
 # MINX
 
@@ -23,20 +22,20 @@ freely, subject to the following restrictions:
 > 3\. This notice may not be removed or altered from any source
 > distribution.
 >
-        */
-#ifndef I_GENERIC_HID_H_
-#define I_GENERIC_HID_H_
+ */
+#ifndef MINX_IGENERICHID_HPP_
+#define MINX_IGENERICHID_HPP_
 
-#include "../API.h"
+#include "../API.hpp"
 
-#include "../GameTime.h"
-#include "../GameComponent.h"
+#include "../GameTime.hpp"
+#include "../GameComponent.hpp"
+#include "../Game.hpp"
+
+#include "Button.hpp"
+#include "Axis.hpp"
 
 #include <GLFW/glfw3.h>
-
-#include "Button.h"
-#include "Axis.h"
-#include "../Game.h"
 #include <vector>
 
 namespace MINX
@@ -55,15 +54,15 @@ namespace MINX
 				 * @param buttonVectorSize The total number of buttons that the device has.
 				 * @param axisVectorSize The total number of axes that the device has.
 				 */
-				IGenericHID(Game * game, unsigned int buttonVectorSize, unsigned int axisVectorSize);
+				IGenericHID(Game* game, unsigned int buttonVectorSize, unsigned int axisVectorSize);
 				
 				/** Contains all of the buttons used by the device.
 				 */
-				vector<Button> * buttons;
+				std::vector<Button>* buttons;
 				
 				/** Contains all of the axes used by the device.
 				 */
-				vector<Axis> * axes;
+				std::vector<Axis>* axes;
 				
 				/** Returns a Button from the specified index in the button vector.
 				 */
