+++ include/MINX/Media/Song.hpp
@@ -23,70 +23,68 @@ freely, subject to the following restrictions:
 > 3\. This notice may not be removed or altered from any source
 > distribution.
 >
-        */
+ */
 
-#ifndef SOUND_FILE_H_
-#define SOUND_FILE_H_
-
-#include "../API.h"
+#include "../API.hpp"
 
 #include "SFML/Audio.hpp"
-#include <string>
+
+#ifndef MINX_SONG_HPP_
+#define MINX_SONG_HPP_
 
 namespace MINX
 {
 	namespace Media
 	{
-		/** A class to hold a single sound file
-		*/
-		class MINX_API SoundFile
+		/** Represents a song loaded from disk.
+		 */
+		class MINX_API Song
 		{
-		public:
-			/* Load a sound file from the hard disk
-			*	@param fileLocation The path of the file to load.
-			*/
-			SoundFile(char* fileLocation);
+			/** Constructs a song from the given file.
+			 */
+			Song(char* fileLocation);
 
-			/* Destroy the sound file safely
-			*/
-			~SoundFile();
+			/** Destroy the sound file safely
+			 */
+			~Song();
 
-			/* unload the sound file safely
-			*/
+			/** unload the sound file safely
+			 */
 			void Unload();
 
-			/* Start playback of the sound file
-			*/
+			/** Start playback of the sound file
+			 */
 			void Play();
 
-			/* Pause playback of the sound file
-			*/
+			/** Pause playback of the sound file
+			 */
 			void Pause();
 
-			/* Stop playback of the sound file
-			*/
+			/** Stop playback of the sound file
+			 */
 			void Stop();
 
-			/* Set the Playback volume of the sound file
-			*	@param volume A double value from 0.0 to 100.0 for the GAIN or volume of the sound file
-			*/
+			/** Set the Playback volume of the sound file
+			 * @param volume A double value from 0.0 to 100.0 for the GAIN or volume of the sound file
+			 */
 			void SetVolume(float volume);
+
+			/** Set whether the loaded song should loop
+			 * @param loop Sets the song to loop if true
+			 */
+			void SetLoop(bool loop);
+
 		private:
 
-			/* The volume of the sound file
-			*/
+			/** The volume of the sound file
+			 */
 			double volume;
-
-			/** The internal sfml sound buffer
+			
+			/** The internal sfml object
 			 */
-			sf::SoundBuffer soundBuffer;
-
-			/** The internal sfml sound object
-			 */
-			sf::Sound sound;
+			sf::Music song;
 		};
 	}
 }
-
 
 #endif
